requirements.txt
flask
mysql-connector
# flask: There are .env or .flaskenv files present. Do "pip install python-dotenv" to use them.
python-dotenv